Warn when trying to build before having run `npm install`

+++ resources/build.sh
@@ -1,5 +1,10 @@
 #!/bin/sh
 
+if [ ! -d "node_modules/.bin" ]; then
+  echo "Be sure to run \`npm install\` before building GraphiQL."
+  exit 1;
+fi
+
 rm -rf dist/ && mkdir -p dist/ &&
 babel src --ignore __tests__ --out-dir dist/ &&
 browserify -g browserify-shim -s GraphiQL dist/index.js > graphiql.js &&
